// vscode/src/generated-sources/modality-api/models/TimelineOverview.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Modality REST API
 * Modality REST API
 *
 * The version of the OpenAPI document: 2.0
 * Contact: support@auxon.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TimelineOverview
 */
export interface TimelineOverview {
    /**
     * 
     * @type {string}
     * @memberof TimelineOverview
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof TimelineOverview
     */
    name: string;
}

/**
 * Check if a given object implements the TimelineOverview interface.
 */
export function instanceOfTimelineOverview(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;

    return isInstance;
}

export function TimelineOverviewFromJSON(json: any): TimelineOverview {
    return TimelineOverviewFromJSONTyped(json, false);
}

export function TimelineOverviewFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimelineOverview {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
    };
}

export function TimelineOverviewToJSON(value?: TimelineOverview | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
    };
}